Add metadata to Logstash events
We have two separate S3 buckets which are setup to notify an SQS queue on object creation. One of these queues is used for ELB access logs and the other is used for S3 access logs. I want to send these logs to the same SQS queue, but doing so would make it impossible to differentiate the different types of logs. To fix this, I have added `[@metadata][s3_bucket_name]` and `[@metadata][s3_object_key]` fields to events that are emitted from the `s3sqs` input plugin. This then allows me to set the type using a filter:

```
filter {
  if [@metadata][s3_bucket_name] == "my-elb-logs-bucket" {
    mutate {
      add_field => {
        "type" => "elb"
      }
    }
  } else if [@metadata][s3_bucket_name] == "my-s3-logs-bucket" {
    mutate {
      add_field => {
        "type" => "s3"
      }
    }
  }
}
```